arreglando modulos cargos, miembros, delegaciones y ciudades
arreglo de mensajes de validacion y de retencion de datos en los
formularios en errores de validacion

+++ database/migrations/2017_03_13_183952_create_telephones_table.php
@@ -15,8 +15,8 @@ class CreateTelephonesTable extends Migration
     {
         Schema::create('telephones', function (Blueprint $table) {
             $table->engine = 'InnoDB';
-            $table->increments('id');
-            $table->integer('member_id');
+            $table->bigIncrements('id');
+            $table->bigInteger('member_id');
             $table->string('telephone', 12);
             $table->timestamps();
         });
