+++ database/migrations/2017_03_13_185538_create_addresses_table.php
@@ -15,10 +15,10 @@ class CreateAddressesTable extends Migration
     {
         Schema::create('addresses', function (Blueprint $table) {
             $table->engine = 'InnoDB';
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('address', 70);
-            $table->integer('city_id');
-            $table->integer('member_id');
+            $table->bigInteger('city_id');
+            $table->bigInteger('member_id');
             $table->timestamps();
         });
     }
